Correccion tipo de dato

+++ src/main/java/com/odonto/dentisys/dto/ProformaDetalleDTO.java
@@ -1,6 +1,5 @@
 package com.odonto.dentisys.dto;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -13,10 +12,10 @@ public class ProformaDetalleDTO {
     private String pacienteNombre;
     private Long medicoId;
     private String medicoNombre;
-    private BigDecimal subtotal;
-    private BigDecimal iva;
-    private BigDecimal descuento;
-    private BigDecimal total;
+    private Double subtotal;
+    private Double iva;
+    private Double descuento;
+    private Double total;
     private String estado;
     private String observaciones;
     private LocalDateTime createdAt;
